Rename relative to path_ext
